Rewrite event loop tests as integration tests
This also removes the duplication of the xk6-timers module, and uses
it directly now that it's an experimental built-in module.

Although, shouldn't these be xk6-timers tests rather than event loop
ones? We have plenty of tests that indirectly test the event loop, and
we could always add much simpler ones that don't involve the timers
module.

See https://github.com/grafana/k6/issues/2459
